tweeked data types to make ids fit in their types and make another id SERIAL

+++ schema.sql
@@ -5,18 +5,18 @@ DROP TABLE IF EXISTS players;
 DROP TABLE IF EXISTS groups;
 
 CREATE TABLE groups (
-    group_id INTEGER PRIMARY KEY,
+    group_id BIGINT PRIMARY KEY,
     group_name VARCHAR(20) NOT NULL
 );
 
 CREATE TABLE games (
-    game_id INTEGER PRIMARY KEY,
+    game_id BIGINT PRIMARY KEY,
     draw_date DATE,
     drawn BOOLEAN DEFAULT FALSE
 );
 
 CREATE TABLE games_and_groups (
-    id INTEGER PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     game_id INTEGER NOT NULL,
     group_id INTEGER NOT NULL,
     FOREIGN KEY(game_id) REFERENCES games(game_id),
@@ -24,7 +24,7 @@ CREATE TABLE games_and_groups (
 );
 
 CREATE TABLE players (
-    player_id INTEGER PRIMARY KEY,
+    player_id BIGINT PRIMARY KEY,
     password_hash VARCHAR(100) NOT NULL,
     player_name VARCHAR(20) NOT NULL,
     picked_name VARCHAR(20) DEFAULT NULL,
